+++ view/src/login/login.css
@@ -3,3 +3,23 @@
     flex-direction: column;
     align-items: center;
 }
+
+.input {
+    width: 100%;
+    padding: 12px 20px;
+    margin: 8px 20px;
+    display: flex;
+    border-radius: 4px;
+    box-sizing: border-box;
+    align-items: center;
+  }
+
+.submitbutton {
+    width: 100%;
+    padding: 12px 20px;
+    margin: 8px 0;
+    display: inline-block;
+    border-radius: 4px;
+    align-items: center;
+}
+
